[objects] Convert usages of FixedArray::length to ulength
Improve robustness in the sandbox attacker model by using unsigned
integer length for FixedArray (see: crrev.com/c/7544875).

Bug: 375937549
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/7580098
Commit-Queue: Arash Kazemi <arashk@chromium.org>
Cr-Commit-Position: refs/heads/main@{#105305}

+++ src/debug/liveedit.cc
@@ -1078,7 +1078,7 @@ void LiveEdit::PatchScript(Isolate* isolate, Handle<Script> script,
     for (Tagged<SharedFunctionInfo> sfi = script_it.Next(); !sfi.is_null();
          sfi = script_it.Next()) {
       DCHECK_EQ(sfi->script(), *new_script);
-      DCHECK_EQ(sfi->function_literal_id(kRelaxedLoad),
+      DCHECK_EQ(static_cast<uint32_t>(sfi->function_literal_id(kRelaxedLoad)),
                 script_it.CurrentIndex());
       // Don't check the start position of the top-level function, as it can
       // overlap with a function in the script.
